Extract Node generation into separate function
